fleming feedback
