+++ notebooks/MAST/JWST/download_by_program_id/requirements.txt
@@ -1,0 +1,2 @@
+astropy >= 5.0.4
+astroquery >= 0.4.6
