--- notebooks/MAST/astroqueryMAST/beginner_program_id_search/requirements.txt
@@ -1,3 +1,0 @@
-astropy >= 5.0.4
-astroquery >= 0.4.6
-numpy >= 1.21.5
